SAMARTHYA TEAM/VIDHATHRI/requirements.txt
python-dotenv
flask
mysql-connector-python
google-generativeai
pymongo